Spam and Non Spam Mail detection
Mails are classified in Spam and Non Spam by studying separate files which has already been labelled as Spam and Non Spam. The training and Test Datasets are provided in separate folders containing Spam Mails and Non Spam Mails and the model is trained accordingly based on the files present in Train folders of each category.